+++ src/components/core/index.ts
@@ -6,3 +6,6 @@ export { Image } from "./Image";
 export { Text } from "./Text";
 export { LineChart } from "./LineChart";
 export { BarChart } from "./BarChart";
+export { ScatterChart } from "./ScatterChart";
+export { PageBreak } from "./PageBreak";
+export { Spacer } from "./Spacer";